// frontend/src/lib/api/models/GlossaryEntryDetail.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { GlossaryEntryRead } from './GlossaryEntryRead';
import type { GlossarySourceReferenceRead } from './GlossarySourceReferenceRead';
export type GlossaryEntryDetail = {
    entry: GlossaryEntryRead;
    sources: Array<GlossarySourceReferenceRead>;
};