// frontend/src/lib/api/models/GlossaryEntryRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GlossaryEntryRead = {
    id: string;
    subjectId: string;
    folderId: (string | null);
    moduleSectionId: (string | null);
    term: string;
    entryType: string;
    language: string;
    shortDefinition: (string | null);
    detailedExplanation: (string | null);
    example: (string | null);
    formulaLatex: (string | null);
    definitionStatus: string;
    status: string;
    createdAt: string;
    updatedAt: string;
};